# doc/api.yml
openapi: 3.0.3
info:
  title: Raibu Backend API
  version: 1.0.0
  description: |
    OpenAPI specification for the Raibu backend.
    This API powers the WebRTC random video chat application, including
    authentication via Google, matchmaking, reporting, and moderation.
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /me:
    get:
      summary: Get current authenticated user
      description: Returns the currently authenticated user (Google OAuth2).
      security:
        - googleOAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          description: Not authenticated
  /sessions/{sessionId}/messages:
    get:
      summary: Get messages for a session
      description: Returns text chat messages for a given session. Only participants of the session can access the messages.
      security:
        - googleOAuth: []
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of chat messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatMessage'
        '401':
          description: Not authenticated
        '403':
          description: Forbidden (not a participant)
        '404':
          description: Session not found
  /sessions/{sessionId}/feedback:
    post:
      summary: Submit feedback for a session
      description: Allows a participant of a session to submit simple like/dislike feedback.
      security:
        - googleOAuth: []
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionFeedbackRequest'
      responses:
        '200':
          description: Feedback recorded (or already present)
        '401':
          description: Not authenticated
        '403':
          description: Forbidden (not a participant)
        '404':
          description: Session not found
...
  /sessions/me:
    get:
      summary: Get current user's chat sessions
      description: Returns chat sessions for the authenticated user.
      security:
        - googleOAuth: []
      responses:
        '200':
          description: List of chat sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatSession'
        '401':
          description: Not authenticated
...
components:
  securitySchemes:
    googleOAuth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/v2/auth
          tokenUrl: https://oauth2.googleapis.com/token
          scopes:
            openid: OpenID Connect scope
            profile: Access to basic profile info
            email: Access to email address
  schemas:
    MeResponse:
      type: object
      properties:
        id:
          type: string
        displayName:
          type: string
        email:
          type: string
          format: email
        avatarUrl:
          type: string
          format: uri
          nullable: true
        countryCode:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
        banned:
          type: boolean
        status:
          $ref: '#/components/schemas/UserStatus'
...
    UserStatus:
      type: string
      enum:
        - IDLE
        - SEARCHING
        - IN_CHAT
        - OFFLINE
...
    ChatSession:
      type: object
      properties:
        id:
          type: string
        user1Id:
          type: string
        user2Id:
          type: string
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
          nullable: true
        status:
          $ref: '#/components/schemas/ChatSessionStatus'
    ChatSessionStatus:
      type: string
      enum:
        - ACTIVE
        - ENDED
...
    ChatMessage:
      type: object
      properties:
        id:
          type: string
        sessionId:
          type: string
        senderId:
          type: string
        content:
          type: string
        sentAt:
          type: string
          format: date-time
    SessionFeedbackRequest:
      type: object
      required:
        - liked
      properties:
        liked:
          type: boolean
...